backend/pyproject.toml
[project]
name = "roboforge"
version = "0.0.1"
description = "RoboForge AI backend — FastAPI service for AI-powered ROS2 development"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "pydantic>=2.10,<3",
    "pyyaml>=6,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["roboforge"]
